test(e2e): tighten network policy pypi probe
